No longer build for IA-64
The COM interface for IA64 has to be rewritten in either two ways: C89
or C++, and what we have now is written in C99. I have no intention to
do it again (I tried, and failed). Probably the easiest way is to use
GCC instead of VS. Therefore disable the build.